refactor: ChatViewController
